requirements-sync.txt
feedparser==6.0.11
html2text==2024.2.26
Pillow==10.4.0
requests==2.32.3
